Print formatted table of issues
This is my own implementation of the table formatting (exercise 4 in
"Organizing a project"). The solution is working properly, however, I
think it could be improved in multiple places.

In #27/#28 I am using an anonymous function for transforming an Integer
into String. If I do not do this, `String.length/1` will fail, since it
is expecting a String and not an Integer. I could not find a method
which does this kind of String interpolation, hence the anonymous
function. Also, I am not sure if you should use the `&` shortcut or a
definition with `fn` in pipe chains. Both look sub-optimal in this
context.

In the print methods I am also using pipes extensively, but only "one
level deep". The anonymous functions passed to `Enum.map/2` could also
utilize the pipe operator, so e.g.
`&(String.pad_trailing(&1, Map.get(col_widths, &1)))` could become
`&(&1 |> String.pad_trailing(Map.get(col_widths, &1)))`. Also the
function could be defined outside of the chain and just referenced.
I am still not sure what is the best practice in cases like these.